Organize tsx files
