+++ src/components/pinContent/swiperPhoto.tsx
