+++ src/components/post/editor.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction, useRef, useEffect } from "react"
+import React, { Dispatch, SetStateAction } from "react"
 import styled from "styled-components"
 import ReactQuill from "react-quill"
 import "react-quill/dist/quill.snow.css"
