+++ src/components/post/uploadPhoto.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { useState, useContext, Dispatch, SetStateAction } from "react"
 import styled from "styled-components"
-import { storage } from "../../Utils/firebase"
-import { AuthContext } from "../../Context/authContext"
+import { storage } from "../../utils/firebase"
+import { AuthContext } from "../../context/authContext"
 import uploadIcon from "../../assets/buttons/uploadImgIcon.png"
 import { ref, uploadBytesResumable, getDownloadURL } from "firebase/storage"
 import imageCompression from "browser-image-compression"
