+++ src/components/profile.tsx
@@ -1,12 +1,11 @@
 import React from "react"
 import { useState, useRef, useContext, useEffect, Dispatch } from "react"
 import styled from "styled-components"
-import { db, storage } from "../Utils/firebase"
+import { db, storage } from "../utils/firebase"
 import imageCompression from "browser-image-compression"
 import { ref, uploadBytesResumable, getDownloadURL } from "firebase/storage"
 import { doc, updateDoc } from "firebase/firestore"
-import { AuthContext } from "../Context/authContext"
-import { Container } from "./styles/mapStyles"
+import { AuthContext } from "../context/authContext"
 import logoutIcon from "../assets/buttons/logoutIcon.png"
 import editPencil from "../assets/buttons/edit.png"
 import spinner from "../assets/dotsSpinner.svg"
