+++ src/components/sampleContent.tsx
@@ -6,7 +6,7 @@ import gala1 from "../assets/samplePhotos/gala1.jpg"
 import gala2 from "../assets/samplePhotos/gala2.jpg"
 import gala3 from "../assets/samplePhotos/gala3.jpg"
 import gala4 from "../assets/samplePhotos/gala4.jpg"
-import { AuthContext } from "../Context/authContext"
+import { AuthContext } from "../context/authContext"
 import mmAvatar from "../assets/samplePhotos/mmAvatar.jpg"
 import pkcAvatar from "../assets/samplePhotos/pkcAvatar.jpg"
 import cyuAvatar from "../assets/samplePhotos/cyuAvatar.jpg"
