+++ src/components/styles/friendStyles.tsx
