+++ src/components/styles/memoriesStyles.tsx
