+++ src/components/styles/modalStyles.tsx
