+++ src/components/styles/widgetStyles.tsx
