+++ src/index.tsx
@@ -4,11 +4,11 @@ import { BrowserRouter, Routes, Route } from "react-router-dom"
 import "./index.css"
 import App from "./App"
 import Home from "./pages/landing"
-import MyMap from "./pages/User/myMap"
-import MyMemories from "./pages/User/myMemories"
-import MyFriends from "./pages/User/myFriends"
-import FriendsMap from "./pages/User/friendMap"
-import FriendMemories from "./pages/User/friendMemories"
+import MyMap from "./pages/user/myMap"
+import MyMemories from "./pages/user/myMemories"
+import MyFriends from "./pages/user/myFriends"
+import FriendsMap from "./pages/user/friendMap"
+import FriendMemories from "./pages/user/friendMemories"
 
 import reportWebVitals from "./reportWebVitals"
