--- src/pages/Utils/gmap.tsx
@@ -1,10 +1,0 @@
-export const myGoogleApiKey = process.env.REACT_APP_google_API_KEY
-
-export const containerStyle = {
-  marginTop: "10px",
-  width: "100%",
-  height: "100%",
-  minHeight: "calc(100vh - 120px)",
-  borderRadius: "10px",
-  zIndex: 30,
-}
